requirements.txt
asgiref==3.3.1
braintree==4.8.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.7
django-debug-toolbar==3.2
django-payments==0.13.0
django-progressive-web-app==0.1.1
django-qurl-templatetag==0.0.14
gunicorn==20.1.0
icalendar==4.0.7
idna==2.10
jsonify==0.5
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
recordtype==1.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.48.0
suds-jurko==0.6
urllib3==1.26.4
urlparse2==1.1.1
whitenoise==5.2.0
xmltodict==0.12.0
